services/transfer/pyproject.toml
[tool.poetry]
name = "transfer-service"
version = "0.1.0"
description = "P2P transfer service between operators"
authors = ["Carpeta Ciudadana Team"]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
asyncpg = "^0.30.0"
psycopg = {extras = ["binary"], version = "^3.1.0"}
httpx = "^0.26.0,<0.29.0"
pydantic = "^2.12.3"
pydantic-settings = "^2.6.0"
tenacity = "^8.2.3"
pyjwt = "^2.8.0"
azure-storage-blob = "^12.19.0"
azure-identity = "^1.15.0"
azure-servicebus = "^7.11.4"
carpeta-common = {path = "../common", develop = true}

[tool.poetry.group.dev.dependencies]
pytest-cov = "^7.0.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
ruff = "^0.1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

